backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydub==0.25.1
mutagen==1.47.0
google-genai==1.0.0
faster-whisper==1.1.1
stable-ts==2.19.1
python-dotenv==1.0.1
aiofiles==24.1.0
websockets==13.0
httpx==0.27.0
Pillow==11.1.0
pydantic-settings==2.5.0
redis==5.0.8
numpy>=1.26.4
librosa>=0.11.0
soundfile>=0.13.1
playwright>=1.40.0
aiohttp>=3.9.0
bcrypt>=4.1.2
email-validator>=2.1.0
# pyjianying -- PyPI 미등록. 필요 시: pip install git+https://github.com/linyiLYi/PyJianYingDraft
